Code refactoring (#5)
